+++ spec/lib/wechat/api_spec.rb
@@ -13,6 +13,14 @@ RSpec.describe Wechat::Api do
     allow(subject.jsapi_ticket).to receive(:jsapi_ticket).and_return('jsapi_ticket')
   end
 
+  describe '#callbackip' do
+    specify 'will get callbackip with access_token' do
+      server_ip_result = 'server_ip_result'
+      expect(subject.client).to receive(:get).with('getcallbackip', params: { access_token: 'access_token' }).and_return(server_ip_result)
+      expect(subject.callbackip).to eq server_ip_result
+    end
+  end
+
   describe '#groups' do
     specify 'will get groups with access_token' do
       groups_result = 'groups_result'